Unify unit conversions into one canonical module
Consolidate the three overlapping converters (utils/unitConversion.ts,
utils/volumeConversion.ts, and units/conversions.ts) into a single canonical
module at packages/lib/src/units/conversions.ts, and delete the two utils
copies. The lib barrel now re-exports from it, so the ~25 cellar/pressing/
packaging/API callers are unchanged.

- Resolves the bushel-factor conflict: standardize on 18.14 kg/bushel (the
  21.7724 path in volumeConversion had zero callers).
- Single VolumeUnit/WeightUnit/formatVolume: formatVolume(liters, unit,
  decimals?) is friendly (whole/1-decimal) without decimals, fixed with —
  satisfying both prior test suites. Standardize on 'mL' (matches additive
  inventory enums); fix volume-input.tsx's lowercase 'ml' outlier.
- bushelsToKg/kgToBushels are now strict (throw on zero, no near-zero clamp),
  fixing 4 long-failing unitConversion tests. Fruit-purchase quantity is
  zod .positive() so this is safe.

Verified: full workspace typecheck clean; lib suite 715/715 green (incl. the
4 fixed tests + canonical module's 39); the 54 failing api integration tests
fail identically on baseline (pre-existing DB/data issues, not this change).
