+++ packages/api/src/routers/index.ts
@@ -95,7 +95,7 @@ import {
   lte,
 } from "drizzle-orm";
 import { TRPCError } from "@trpc/server";
-import { convertVolume, roundToDecimals } from "lib/src/utils/volumeConversion";
+import { convertVolume, roundToDecimals } from "lib/src/units/conversions";
 import {
   publishCreateEvent,
   publishUpdateEvent,
